Clean up obsolete classes, remove dependency on net.lingala.zip4j from commons to app and drop obsolete commons dependency on org.springframework:spring-web
